Move molecule components with their stories into seperate directories
