+++ .vscode/settings.json
@@ -25,7 +25,9 @@
 		"InLink": "Cluster",
 		"NavLink": "Cluster",
 		"SkipLink": "Cluster",
-		"02_molecules": "App",
+        "02_molecules": "App",
+        "MenuDesktop": "App",
+		"MenuMobile": "App",
 		"03_organisms": "Components",
 		"04_templates": "Layout",
 		"05_pages": "Log",
