+++ src/02_molecules/MenuDesktop/MenuDesktop.js
@@ -6,7 +6,7 @@
  */
 
 import React from "react"
-import NavLink from "../01_atoms/NavLink/NavLink"
+import NavLink from "../../01_atoms/NavLink/NavLink"
 
 const MenuDesktop = () => {
     return (
