+++ src/03_organisms/Header.js
@@ -8,8 +8,8 @@
 import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
 import InLink from "../01_atoms/InLink/InLink"
-import MenuMobile from "../02_molecules/MenuMobile"
-import MenuDesktop from "../02_molecules/MenuDesktop"
+import MenuMobile from "../02_molecules/MenuMobile/MenuMobile"
+import MenuDesktop from "../02_molecules/MenuDesktop/MenuDesktop"
 
 const Header = () => {
     const data = useStaticQuery(graphql`
